# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalog API",
        "description": "JSON API for catalog search and categories.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current application host"
        }
    ],
    "paths": {
...
        "/api/catalog/search": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "operationId": "catalogSearch",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 100
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 20,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search hits or empty list when query is missing or shorter than 2 characters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "data",
                                        "meta"
                                    ],
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CatalogProduct"
                                            }
                                        },
                                        "meta": {
                                            "required": [
                                                "limit",
                                                "query"
                                            ],
                                            "properties": {
                                                "limit": {
                                                    "type": "integer"
                                                },
                                                "query": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ApiError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string",
                        "nullable": true
                    },
                    "errors": {
                        "description": "Validation errors keyed by field name",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CatalogProduct": {
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 100
                    },
                    "title": {
                        "type": "string",
                        "example": "Product"
                    },
                    "price": {
                        "description": "Numeric price or string from upstream catalog",
                        "nullable": true,
                        "oneOf": [
                            {
                                "type": "number",
                                "format": "float"
                            },
                            {
                                "type": "string"
                            }
                        ]
                    },
                    "image_url": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}